Update to upstream.

+++ src/lily_parser.h
@@ -2,17 +2,15 @@
 # define LILY_PARSER_H
 
 # include "lily.h"
-
-# include "lily_raiser.h"
-# include "lily_expr.h"
-# include "lily_lexer.h"
-# include "lily_emitter.h"
-# include "lily_symtab.h"
-# include "lily_vm.h"
-# include "lily_type_maker.h"
 # include "lily_buffer_u16.h"
-# include "lily_value_stack.h"
+# include "lily_emitter.h"
+# include "lily_expr.h"
 # include "lily_generic_pool.h"
+# include "lily_lexer.h"
+# include "lily_raiser.h"
+# include "lily_symtab.h"
+# include "lily_type_maker.h"
+# include "lily_vm.h"
 
 /* This is used to prevent multiple content loads and to make sure content
    handling has content. */
@@ -93,7 +91,7 @@ typedef struct lily_parse_state_ {
     lily_doc_stack *doc;
 } lily_parse_state;
 
-lily_var *lily_parser_lambda_eval(lily_parse_state *, int, const char *,
+lily_var *lily_parser_lambda_eval(lily_parse_state *, uint16_t, const char *,
         lily_type *);
 lily_item *lily_find_or_dl_member(lily_parse_state *, lily_class *,
         const char *);
